requirements.txt
python-fasthtml
uvicorn>=0.29
autopep8
